# swagger-doc.yaml
openapi: 3.0.3
info:
  title: Work flex - OpenAPI 3.0
  contact:
    email: lmnhut1612@gmail.com
  version: 1.0.1
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /permission:
    post:
      tags:
        - permission
      summary: Thêm mới phân quyền
      description: Thêm phân quyền mới vào database
      security:
        - bearerAuth: []
      operationId: addPermission
      parameters:
        - in: header
          name: x-client-id
          schema:
            type: string
            format: uuid
          required: true
          description: Client ID for authentication
      requestBody:
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                  example: add-permission
                description:
                  type: string
                  example: Cho phép thêm phân quyền cho hệ thống
        required: true
      responses:
        '201':
          description: Thêm mới thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      metadata:
                        type: object
                        properties:
                          permission:
                            type: object
                            properties:
                              title:
                                type: string
                                example: add-permission
                              description:
                                type: string
                                example: Cho phép thêm phân quyền cho hệ thống
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: UnprocessableEntity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          description: Thông báo về kết quả của operation
        status:
          type: integer
          example: 2xx
          description: Mã trạng thái HTTP
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          description: Trạng thái của response (thường là "fail")
        message:
          type: string
          description: Thông báo lỗi chi tiết

  # requestBodies:
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: x-api-key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT